+++ src/rules/mod.rs
@@ -20,12 +20,31 @@ pub fn builtin_rules(config: Option<&Config>) -> Vec<Box<dyn Rule>> {
         }) => r,
         _ => &default_rules_config,
     };
+    let excludes = rules_config.exclude.clone().unwrap_or_default();
 
+    let mut rules = Vec::new();
     // Each Rule uses inventory::submit! to register a factory to build themselves with.
-    inventory::iter::<RuleFactory>
-        .into_iter()
-        .map(|factory| factory.build(rules_config))
-        .collect()
+    'outer: for factory in inventory::iter::<RuleFactory>.into_iter() {
+        let rule = factory.build(rules_config);
+        let string_id = rule.id().to_string();
+        let short_pretty_id = format!("H{}", rule.id());
+        let pretty_id = rule.pretty_id(); // each call does an allocation
+
+        for exclude in &excludes {
+            if &string_id == exclude
+                || &short_pretty_id == exclude
+                || rule.human_id() == exclude
+                || &pretty_id == exclude
+            {
+                tracing::info!("Excluding rule: {pretty_id}");
+                continue 'outer;
+            }
+        }
+
+        rules.push(rule);
+    }
+
+    rules
 }
 
 pub fn builtin_rules_all() -> Vec<Box<dyn Rule>> {
@@ -105,5 +124,28 @@ mod tests {
             assert_ne!(rule.name().to_lowercase(), "todo");
             assert_ne!(rule.description().to_lowercase(), "todo");
         }
+    }
+
+    #[test]
+    fn exclude_rules() {
+        let config = RulesConfig {
+            exclude: Some(vec![
+                "1".to_string(),
+                "H2".to_string(),
+                "longest-subject-line".to_string(),
+                "H4-non-unicode".to_string(),
+            ]),
+            ..Default::default()
+        };
+        let config = Config {
+            rules: Some(config),
+            ..Default::default()
+        };
+        let rules = builtin_rules(Some(&config));
+        let ids: Vec<_> = rules.iter().map(|r| r.id()).collect();
+        assert!(!ids.contains(&1));
+        assert!(!ids.contains(&2));
+        assert!(!ids.contains(&3));
+        assert!(!ids.contains(&4));
     }
 }